django_bp_system/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
Pillow>=10.0.0